Start implementation session manager

+++ unity/spirit_m2m_webrtc/Assets/Scripts/Runtime/MultiCapture/Test/MultiCapturingTest.cs
@@ -75,7 +75,7 @@ public class MultiCapturingTest : MonoBehaviour
         DecodedPointCloudData pcData;
         if (!inProgessFrames.TryGetValue(frameNr, out pcData))
         {
-            pcData = new DecodedPointCloudData((int)frameNr, 125000, 3, new List<bool> { true, true, true }, timestamp)  ;
+            pcData = new DecodedPointCloudData(frameNr, 125000, 3, new List<bool> { true, true, true }, timestamp)  ;
             inProgessFrames.Add(frameNr, pcData);
         }
         unsafe
