+++ unity/spirit_m2m_webrtc/Assets/Scripts/Runtime/PointCloud/Test/CapturingTestMultiAudio.cs
@@ -79,7 +79,7 @@ public class CapturingTestMultiAudio : MonoBehaviour
         DecodedPointCloudData pcData;
         if (!inProgessFrames.TryGetValue(frameNr, out pcData))
         {
-            pcData = new DecodedPointCloudData((int)frameNr, 125000, 3, new List<bool> { true, true, true }, timestamp)  ;
+            pcData = new DecodedPointCloudData(frameNr, 125000, 3, new List<bool> { true, true, true }, timestamp)  ;
             inProgessFrames.Add(frameNr, pcData);
         }
         unsafe
